+++ src/core/index.js
@@ -108,6 +108,7 @@ import redo from './redo';
 import updateAnnotationState from './updateAnnotationState';
 import getFontStyles from './getFontStyles';
 import mergeDocument from './mergeDocument';
+import extractPagesWithAnnotations from './extractPagesWithAnnotations';
 
 export default {
   setToolMode,
@@ -221,4 +222,5 @@ export default {
   updateAnnotationState,
   getFontStyles,
   mergeDocument,
+  extractPagesWithAnnotations,
 };
